fix: add useCanvasActions
