+++ packages/nextjs/app/newcreate/page.tsx
@@ -8,6 +8,7 @@ import { CanvasActions } from "./_components/CanvasActions";
 import { CanvasControls } from "./_components/CanvasControls";
 import { CreateInkForm } from "./_components/CreateInkForm";
 import { DraftManager } from "./_components/DraftManager";
+import { useCanvasActions } from "./_hooks/useCanvasActions";
 import { useCreateInk } from "./_hooks/useCreateInk";
 import "./styles.css";
 import LZ from "lz-string";
@@ -196,33 +197,12 @@ const CreateInk = () => {
     saveDrawing(drawingCanvas.current, true);
   };
 
-  const undo = () => {
-    if (!drawingCanvas?.current?.lines?.length) return;
-
-    if (drawingCanvas.current.lines[drawingCanvas.current.lines.length - 1]?.ref) {
-      drawingCanvas.current.lines[0].brushColor =
-        drawingCanvas.current.lines[drawingCanvas.current.lines.length - 1].brushColor;
-      const lines = drawingCanvas.current.lines.slice(0, -1);
-      triggerOnChange(lines);
-    } else {
-      const lines = drawingCanvas.current.lines.slice(0, -1);
-      triggerOnChange(lines);
-    }
-  };
-
-  const downloadCanvas = async () => {
-    const myData = drawingCanvas?.current?.getSaveData();
-    const fileName = `nifty_ink_canvas_${Date.now()}`;
-    const json = JSON.stringify(myData);
-    const blob = new Blob([json], { type: "application/json" });
-    const href = await URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = href;
-    link.download = fileName + ".json";
-    document.body.appendChild(link);
-    link.click();
-    document.body.removeChild(link);
-  };
+  const { undo, downloadCanvas, uploadCanvas, fillBackground, drawFrame, saveDraft } = useCanvasActions(
+    drawingCanvas,
+    triggerOnChange,
+    setDrafts,
+    setCanvasFile,
+  );
 
   const uploadFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const fileReader = new FileReader();
@@ -233,88 +213,6 @@ const CreateInk = () => {
   };
 
   const uploadRef = useRef<HTMLInputElement | null>(null);
-
-  const uploadCanvas = (uploadedDrawing: any) => {
-    drawingCanvas?.current?.loadSaveData(uploadedDrawing);
-    saveDrawing(drawingCanvas.current, true);
-    setCanvasFile(undefined);
-  };
-
-  const fillBackground = (color: string) => {
-    if (!drawingCanvas.current) return;
-    const width = drawingCanvas?.current.props.canvasWidth;
-    const height = drawingCanvas?.current.props.canvasHeight;
-
-    const bg = {
-      brushColor: color,
-      brushRadius: (width + height) / 2,
-      points: [
-        { x: 0, y: 0 },
-        { x: width, y: height },
-      ],
-      background: true,
-    };
-
-    const previousBGColor = drawingCanvas.current.lines.filter(l => l?.ref).length
-      ? drawingCanvas.current.lines[0].brushColor
-      : "#FFF";
-
-    const bgRef = {
-      brushColor: previousBGColor,
-      brushRadius: 1,
-      points: [
-        { x: -1, y: -1 },
-        { x: -1, y: -1 },
-      ],
-      ref: true,
-    };
-
-    drawingCanvas.current.lines.filter(l => l.background).length
-      ? drawingCanvas.current.lines.splice(0, 1, bg)
-      : drawingCanvas.current.lines.unshift(bg);
-    drawingCanvas.current.lines.push(bgRef);
-
-    const lines = drawingCanvas.current.lines;
-
-    triggerOnChange(lines);
-  };
-
-  const drawFrame = (color: string, radius: number) => {
-    if (!drawingCanvas.current) return;
-
-    const width = drawingCanvas.current.props.canvasWidth;
-    const height = drawingCanvas.current.props.canvasHeight;
-
-    drawingCanvas.current.lines.push({
-      brushColor: color,
-      brushRadius: radius,
-      points: [
-        { x: 0, y: 0 },
-        { x: width, y: 0 },
-        { x: width, y: 0 },
-        { x: width, y: height },
-        { x: width, y: height },
-        { x: 0, y: height },
-        { x: 0, y: height },
-        { x: 0, y: 0 },
-      ],
-    });
-
-    const lines = drawingCanvas.current.lines;
-
-    triggerOnChange(lines);
-  };
-
-  const saveDraft = () => {
-    if (!drawingCanvas.current) return;
-
-    const imageData = drawingCanvas?.current?.canvas?.drawing.toDataURL("image/png");
-    const savedData = LZ.compress(drawingCanvas.current.getSaveData());
-
-    setDrafts((drafts: any) => {
-      return [...drafts, { imageData, savedData }];
-    });
-  };
 
   useEffect(() => {
     if (isSaving) {
@@ -380,7 +278,6 @@ const CreateInk = () => {
             onTouchEnd={debouncedSaveCanvas}
           >
             <CanvasDraw
-              // key={mode + "" + canvasKey}
               ref={drawingCanvas}
               canvasWidth={size[0]}
               canvasHeight={size[1]}
